Refractor about half done

+++ source/shaders.c
@@ -5,25 +5,31 @@
 #include <fcntl.h>
 #include <sys/stat.h>
 #include <sys/types.h>
+#include <stddef.h>
 #include "shaders.h"
 
 /* Attempts to load a file into a text buffer.
- * Will return null if runs into an exception loading the 
+ * Will return null if runs into an exception loading the
  * file.
  */
 const char* load_file_into_buffer(const char* file_path) {
     int fd = open(file_path, O_RDONLY);
     if(fd < 0)
         return NULL;
+
     //Find the size of the file.
     struct stat stat_struct;
     if(fstat(fd, &stat_struct) < 0)
         return NULL;
 
+    //Allocate a one longer buffer.
     char* buff = malloc(stat_struct.st_size + 1);
+
+    //Read the entire file into this buffer and null terminate it.
     read(fd, buff, stat_struct.st_size);
     *(buff + stat_struct.st_size) = '\0';
     close(fd);
+    
     return buff;
 }
 
@@ -55,10 +61,10 @@ GLuint init_shader(const char* shader_path, GLenum shader_type) {
     return shader;
 }
 
-GLuint init_program(GLuint shaders[], int shaders_num) {
+GLuint init_program(GLuint shaders[], size_t shaders_num) {
     GLuint program = glCreateProgram();
 
-    for(int i = 0; i < shaders_num; i++)
+    for(size_t i = 0; i < shaders_num; i++)
         glAttachShader(program, shaders[i]);
 
     glLinkProgram(program);
@@ -68,13 +74,13 @@ GLuint init_program(GLuint shaders[], int shaders_num) {
     if(stat == GL_FALSE) {
         GLint info_len;
         glGetProgramiv(program, GL_INFO_LOG_LENGTH, &info_len);
-        
+
         GLchar info_log[info_len + 1];
         glGetProgramInfoLog(program, info_len, NULL, info_log);
         fprintf(stderr, "Shader Program Linking Failure!\nError:%s\n", info_log);
     }
 
-    for(int i = 0; i < shaders_num; i++)
+    for(size_t i = 0; i < shaders_num; i++)
         glDetachShader(program, shaders[i]);
 
     return program;
