+++ source/window.c
@@ -9,7 +9,7 @@ void init_glut(int *argc, char **argv) {
     glutInit(argc, argv);
 }
 
-void init_window(int width, int height) {
+void init_window(uint16_t width, uint16_t height) {
     glutInitContextVersion(3, 3);
     glutInitWindowPosition(0, 0);
     glutInitDisplayMode(GLUT_RGB | GLUT_DOUBLE);
